应该使用反射注册
