+++ gradle.properties
@@ -2,7 +2,7 @@ org.gradle.jvmargs=-Xmx1G
 
 # Project settings
 # Rule: [major update].[feature update].[bug fix]
-project_version=0.0.47.2
+project_version=0.0.47.3
 config_version=25
 lang_version=4
 project_group=net.momirealms
